Add init file to bill_crawler
