+++ runner_pipelines/bill_votes_df.py
@@ -1,6 +1,6 @@
 """
 
-This script is a pipeline built in order to record all of the bill cosponsors to any presented bills. We will receive
+This script is a pipeline built in order to record all of the bill votes from the House and Senate. We will be receiving
 these votes from a Scrapy pipeline that has sent this data to PubSub. This pipeline receives the data from PubSub,
 formats and filters the data, and then sends along the clean data to BQ for storage.
 
@@ -21,6 +21,14 @@ import apache_beam as beam
 from apache_beam.options.pipeline_options import PipelineOptions
 from apache_beam.io.gcp.pubsub import ReadFromPubSub
 
+# Date modules for date formatting
+from datetime import date
+
+# Miscellaneous modules
+import re
+import unidecode
+from datetime import date, timedelta
+
 # Import transforms file
 import transforms.df_ptransforms as pt
 
@@ -30,16 +38,17 @@ logging.basicConfig(level=logging.INFO)
 
 # Location of the PubSub topic and subscription we will be using to send/receive messages
 project_id = 'politics-data-tracker-1'
-topic_name = 'cosponsors'
-subscription_name = 'cosponsors_df'
+topic_name = 'bill_votes'
+subscription_name = 'test_bill_votes_sub'
 
 # Locations in BigQuery that this script uses to read/write.
-dataset_id = 'poliviews'
-cosponsors_table_id = 'bill_cosponsors'
-nickname_table_id = 'nicknames'
+dataset_id = 'runner_pipelines'
+dest_table_id = 'all_votes' # Location for writing any voting information
+nickname_table_id = 'nicknames' # Location to read list of any nicknames in order to properly edit politician info.
+house_tbl_id = 'house'
 
 # PubSub will only accept long strings as names of exact locations of BigQuery tables.
-bill_spec = '{0}:{1}.{2}'.format(project_id, dataset_id, cosponsors_table_id)
+bq_spec = '{0}:{1}.{2}'.format(project_id, dataset_id, dest_table_id)
 nickname_spec = '{0}:{1}.{2}'.format(project_id, dataset_id, nickname_table_id)
 
 # Also included is the current name of the script
@@ -48,62 +57,39 @@ script_name = os.path.basename(sys.argv[0])
 # This is a list of all attributes that will be present in either a good vote or error vote. This will be used
 # to filter and/or order the attributes accordingly.
 # At this time, this is necessary. Make sure to do this for all pipelines.
-attributes_lst = ['bill_id','amdt_id','first_name','last_name','party','state']
+attributes_lst = ['bill_id','amdt_id','first_name','last_name','party','state','vote_cast','vote_date','chamber']
 error_attr_lst = ['error_msg', 'ptransform', 'script']
 full_lst = attributes_lst + error_attr_lst
 
+# Set date for yesterday's bills that are published
+date_yesterday = date.today() - timedelta(days=1)
+this_year = date_yesterday.year
+
 # Set all options needed to properly run the pipeline. This pipeline will run on Dataflow as a streaming pipeline.
-options = PipelineOptions(
-    streaming=False,
-    runner='DirectRunner',
-    project=project_id,
-    temp_location='gs://{0}/tmp'.format(project_id),
-    staging_location='gs://{0}/staging'.format(project_id))
+options = PipelineOptions(streaming=False,
+                          runner='DirectRunner',
+                          project=project_id,
+                          temp_location='gs://{0}/tmp'.format(project_id),
+                          staging_location='gs://{0}/staging'.format(project_id))
 
 # This builds the Beam pipeline in order to run Dataflow
 p = beam.Pipeline(options = options)
 logging.info('Created Dataflow pipeline.')
 
-class NormalizeAttributesFn(beam.DoFn):
-    def process(self, element):
-        try:
-            element['first_name']=element.pop('cosponsor_fn')
-            element['last_name']=element.pop('cosponsor_ln')
-            element['party']=element.pop('cosponsor_party')
-            element['state']=element.pop('cosponsor_state')
-            yield element
-        except Exception as e:
-            element['error_msg'] = e
-            element['ptransform'] = self.__class__.__name__
-            element['error_tag'] = True
-            element['script'] = os.path.basename(sys.argv[0])
-            logging.info('Error received at {0}: {1}'.format(self.__class__.__name__, element))
-            yield element
-
-class RevertAttributesFn(beam.DoFn):
-    def process(self, element):
-        try:
-            element['cosponsor_fn']=element.pop('first_name')
-            element['cosponsor_ln']=element.pop('last_name')
-            element['cosponsor_party']=element.pop('party')
-            element['cosponsor_state']=element.pop('state')
-            yield element
-        except Exception as e:
-            element['error_msg'] = e
-            element['ptransform'] = self.__class__.__name__
-            element['error_tag'] = True
-            element['script'] = os.path.basename(sys.argv[0])
-            logging.info('Error received at {0}: {1}'.format(self.__class__.__name__, element))
-            yield element
-
 # This will pull in all of the recorded nicknames to compare to the incoming PubSubMessages. This is needed to filter
 # and normalize the data.
 client = bigquery.Client()
-query_job = client.query("""
+nickname_query = client.query("""
     select * from `{0}.{1}.{2}`""".format(project_id, dataset_id, nickname_table_id))
-nickname_tbl = query_job.result()
+nickname_tbl = nickname_query.result()
 nickname_tbl = [dict(row.items()) for row in nickname_tbl]
 nickname_tbl = [{str(k):str(v) for (k,v) in d.items()} for d in nickname_tbl]
+
+house_query = client.query("""
+    select * from `{0}.{1}.{2}` where date < {3}""".format(project_id, dataset_id, house_tbl_id, date_yesterday))
+house_tbl = house_query.result()
+house_tbl = [dict(row.items()) for row in house_tbl]
+house_tbl = [{str(k):str(v) for (k,v) in d.items()} for d in house_tbl]
 
 class SplitFn(beam.DoFn):
     def process(self, element):
@@ -111,47 +97,60 @@ class SplitFn(beam.DoFn):
             index, \
             bill_id, \
             amdt_id, \
-            cosponsor_fn, \
-            cosponsor_ln, \
-            cosponsor_party, \
-            cosponsor_state = element.split(',')
+            first_name, \
+            last_name, \
+            party, \
+            state, \
+            vote_cast, \
+            vote_date, \
+            chamber, \
+            chamber_state = element.split(',')
 
             d = {
                 'bill_id': bill_id,
                 'amdt_id': amdt_id,
-                'cosponsor_fn': cosponsor_fn,
-                'cosponsor_ln': cosponsor_ln,
-                'cosponsor_party': cosponsor_party,
-                'cosponsor_state': cosponsor_state
+                'first_name': first_name,
+                'last_name': last_name,
+                'party': party,
+                'state': state,
+                'vote_cast': vote_cast,
+                'vote_date': vote_date,
+                'chamber': chamber,
+                'chamber_state': chamber_state
             }
             yield d
 
-cosponsor = (
+# Runs the main part of the pipeline. Errors will be tagged, good votes will continue on to BQ.
+vote = (
     p
-    | 'Read from CSV' >> beam.io.ReadFromText('{0}/tmp/cosponsors/*.csv'.format(os.path.expanduser('~')),
+    | 'Read from CSV' >> beam.io.ReadFromText('{0}/tmp/bill_votes/*.csv'.format(os.path.expanduser('~')),
                                               skip_header_lines=1)
     | 'Split Values' >> beam.ParDo(SplitFn())
-    | 'Normalize Attributes' >> beam.ParDo(NormalizeAttributesFn())
-    | 'Scrub First Name' >> beam.ParDo(pt.ScrubFnameFn(), keep_suffix=False)
-    | 'Fix Nicknames' >> beam.ParDo(pt.FixNicknameFn(), n_tbl = nickname_tbl, keep_nickname=False)
+    | 'Isolate Attributes' >> beam.ParDo(pt.IsolateAttrFn())
+    | 'Fix Value Types' >> beam.ParDo(pt.FixTypesFn(), int_lst=['vote_cast'])
+    | 'Fix House First Names' >> beam.ParDo(pt.FixHouseFirstNameFn(), tbl=house_tbl)
+    | 'Scrub First Name' >> beam.ParDo(pt.ScrubFnameFn())
+    | 'Fix Nicknames' >> beam.ParDo(pt.FixNicknameFn(), n_tbl=nickname_tbl)
     | 'Scrub Last Name' >> beam.ParDo(pt.ScrubLnameFn())
-    # | 'Revert Attributes' >> beam.ParDo(RevertAttributesFn())
     | 'Fix Nones' >> beam.ParDo(pt.FixNoneFn())
     | 'Tag Errors' >> beam.ParDo(pt.TagErrorsFn()).with_outputs('error_tag'))
 
 # Separates the clean and error votes for proper processing.
-error_cosponsors = cosponsor.error_tag
-clean_cosponsors = cosponsor[None]
+error_votes = vote.error_tag
+clean_votes = vote[None]
 
-(clean_cosponsors
+# The cleaned elements will be sent to the proper BQ table for storage.
+(clean_votes
     | 'Filter Keys' >> beam.ParDo(pt.FilterKeysFn(), attr_lst=attributes_lst)
     | 'Write to BQ' >> beam.io.WriteToBigQuery(
-        table=bill_spec,
-        write_disposition=beam.io.BigQueryDisposition.WRITE_APPEND,
-        create_disposition=beam.io.BigQueryDisposition.CREATE_NEVER
+        table = bq_spec,
+        write_disposition = beam.io.BigQueryDisposition.WRITE_APPEND,
+        create_disposition = beam.io.BigQueryDisposition.CREATE_NEVER
 ))
 
-(error_cosponsors
+# This pipeline will take all of the error votes, clean them, properly orders and formats them,
+# And outputs them into a friendly CSV format.
+(error_votes
     | 'Make All Strings' >> beam.ParDo(pt.MakeAllStringsFn())
     | 'CSV Formatting' >> beam.ParDo(pt.BuildCSVRowFn(), lst=full_lst)
     | 'Write to CSV' >> beam.io.WriteToText(
