+++ runner_pipelines/senate_members_df.py
@@ -1,6 +1,6 @@
 """
 
-This script is a pipeline built in order to record all of the politicians currently in the House. We will be receiving
+This script is a pipeline built in order to record all of the politicians currently in the Senate. We will be receiving
 these votes from a Scrapy pipeline that has sent this data to PubSub. This pipeline receives the data from PubSub,
 formats and filters the data, and then sends along the clean data to BQ for storage.
 
@@ -23,7 +23,6 @@ from apache_beam.io.gcp.pubsub import ReadFromPubSub
 
 # Import transforms file
 import transforms.df_ptransforms as pt
-import pandas as pd
 
 # This module will closely track the progression of the PCollection through the pipeline.
 # It has been set to report on the collection throughout its creation and augmentation.
@@ -31,18 +30,18 @@ logging.basicConfig(level=logging.INFO)
 
 # Location of the PubSub topic and subscription we will be using to send/receive messages
 project_id = 'politics-data-tracker-1'
-topic_name = 'house_pols'
-subscription_name = 'house_pols_df'
+topic_name = 'senate_pols'
+subscription_name = 'senate_pols_df'
 
 # Locations in BigQuery that this script uses to read/write.
-dataset_id = 'poliviews'
-house_table_id = 'house' # Location for writing all Rep info
+dataset_id = 'runner_pipelines'
+senate_table_id = 'senate' # Location for writing all senator info
 pol_table_id = 'politicians' # Location for writing any new politicians
 nickname_table_id = 'nicknames' # Location to read list of any nicknames in order to properly edit politician info.
 
 # PubSub will only accept long strings as names of exact locations of BigQuery tables.
 pol_spec = '{0}:{1}.{2}'.format(project_id, dataset_id, pol_table_id)
-house_spec = '{0}:{1}.{2}'.format(project_id, dataset_id, house_table_id)
+senate_spec = '{0}:{1}.{2}'.format(project_id, dataset_id, senate_table_id)
 
 # Also included is the current name of the script
 script_name = os.path.basename(sys.argv[0])
@@ -51,7 +50,7 @@ script_name = os.path.basename(sys.argv[0])
 # to filter and/or order the attributes accordingly.
 # At this time, this is necessary. Make sure to do this for all pipelines.
 pol_attr_lst = ['first_name','last_name','party','state', 'suffix', 'nickname']
-house_attr_lst = ['first_name', 'last_name', 'party', 'state', 'district', 'date']
+senate_attr_lst = ['first_name', 'last_name', 'party', 'state', 'date']
 error_attr_lst = ['error_msg', 'ptransform', 'script']
 pol_full_lst = pol_attr_lst + error_attr_lst
 
@@ -64,7 +63,7 @@ options = PipelineOptions(streaming=False,
 
 # This builds the Beam pipeline in order to run Dataflow
 p = beam.Pipeline(options=options)
-logging.info('Created Beam pipeline.')
+logging.info('Created Dataflow pipeline.')
 
 # This will pull in all of the recorded nicknames to compare to the incoming PubSubMessages. This is needed to filter
 # and normalize the data.
@@ -83,61 +82,29 @@ pols_tbl = pols_query.result()
 pols_tbl = [dict(row.items()) for row in pols_tbl]
 pols_tbl = [{str(k):str(v) for (k,v) in d.items()} for d in pols_tbl]
 
-# The House elements pull states as full names instead of initials.
-# This query as well as the following DoFn are meant to account for that and map the initals back to the proper states.
-state_query = client.query("""
-    select * from `{0}.{1}.{2}`""".format(project_id, dataset_id, 'state_map'))
-state_tbl = state_query.result()
-state_tbl = [dict(row.items()) for row in state_tbl]
-state_tbl = [{str(k):str(v) for (k,v) in d.items()} for d in state_tbl]
-
-class StateMapFn(beam.DoFn):
-    def process(self, element, tbl):
-        try:
-            import pandas as pd
-            tbl = pd.DataFrame(tbl)
-            tbl_matched = tbl[tbl['state']==element['state']]
-            if len(tbl_matched.index) > 1:
-                raise ValueError('There are multiple states in the database that match the given state.')
-            elif len(tbl_matched.index) == 1:
-                element['state'] = tbl_matched['initials'].values[0]
-                yield element
-            else:
-                raise ValueError('There are no states in the database that match the given state.')
-        except Exception as e:
-            element['error_msg'] = e
-            element['ptransform'] = self.__class__.__name__
-            element['error_tag'] = True
-            element['script'] = os.path.basename(sys.argv[0])
-            logging.info('Error received at {0}: {1}'.format(self.__class__.__name__, element))
-            yield element
-        yield element
-
 class SplitFn(beam.DoFn):
     def process(self, element):
         # for i in element:
             logging.info('{0}: {1}'.format(self.__class__.__name__, element))
-            index, first_name, last_name, party, state, district = element.split(',')
+            index, first_name, last_name, party, state = element.split(',')
             d = {
                 'first_name': first_name,
                 'last_name': last_name,
                 'party': party,
-                'state': state,
-                'district': district
+                'state': state
             }
             yield d
 
 # Runs the main part of the pipeline. Errors will be tagged, clean politicians will continue on to BQ.
 pol = (
         p
-        | 'Read from CSV' >> beam.io.ReadFromText('{0}/tmp/house_members/*.csv'.format(os.path.expanduser('~')),
+        | 'Read from CSV' >> beam.io.ReadFromText('{0}/tmp/senate_members/*.csv'.format(os.path.expanduser('~')),
                                                   skip_header_lines=1)
         | 'Split Values' >> beam.ParDo(SplitFn())
         # | 'Isolate Attributes' >> beam.ParDo(pt.IsolateAttrFn())
         | 'Scrub First Name' >> beam.ParDo(pt.ScrubFnameFn(), keep_suffix=True)
         | 'Fix Nicknames' >> beam.ParDo(pt.FixNicknameFn(), n_tbl=nickname_tbl, keep_nickname=True)
         | 'Scrub Last Name' >> beam.ParDo(pt.ScrubLnameFn())
-        | 'Map States' >> beam.ParDo(StateMapFn(), tbl=state_tbl)
         | 'Fix Nones' >> beam.ParDo(pt.FixNoneFn())
         | 'Tag Errors' >> beam.ParDo(pt.TagErrorsFn()).with_outputs('error_tag'))
 
@@ -146,27 +113,27 @@ clean_pols = pol[None]
 
 # A new Politician will only be published if thy are not already contained in the Politicians table. If they are new,
 # they will be properly uploaded. If they are not new, then they will be ignored in this pipeline.
-new_pol = (clean_pols
-           | 'Filter Existing Pols' >> beam.ParDo(pt.NewPolsOnlyFn(), pol_tbl=pols_tbl)
-           | 'Filter Keys' >> beam.ParDo(pt.FilterKeysFn(), attr_lst=pol_attr_lst)
-           | 'Write Pol to BQ' >> beam.io.WriteToBigQuery(
+new_pol = (
+        clean_pols
+        | 'Filter Existing Pols' >> beam.ParDo(pt.NewPolsOnlyFn(), pol_tbl=pols_tbl)
+        | 'Filter Pol Keys' >> beam.ParDo(pt.FilterKeysFn(), attr_lst=pol_attr_lst)
+        | 'Write Pol to BQ' >> beam.io.WriteToBigQuery(
             table=pol_spec,
             write_disposition=beam.io.BigQueryDisposition.WRITE_APPEND,
             create_disposition=beam.io.BigQueryDisposition.CREATE_NEVER
         ))
 
-# A new Rep line will be published every day in this pipeline. We add today's date to record their time in office
+# A new Senator line will be published every day in this pipeline. We add today's date to record their time in office
 # and will upload a row signifying their time in office on that day.
-new_rep = (
+new_senator = (
         clean_pols
         | 'Add Date' >> beam.ParDo(pt.AddDateFn())
-        | 'Filter House Keys' >> beam.ParDo(pt.FilterKeysFn(), attr_lst=house_attr_lst)
-        | 'Write Rep to BQ' >> beam.io.WriteToBigQuery(
-            table=house_spec,
+        | 'Filter Sen Keys' >> beam.ParDo(pt.FilterKeysFn(), attr_lst=senate_attr_lst)
+        | 'Write Sen to BQ' >> beam.io.WriteToBigQuery(
+            table=senate_spec,
             write_disposition=beam.io.BigQueryDisposition.WRITE_APPEND,
             create_disposition=beam.io.BigQueryDisposition.CREATE_NEVER
         ))
-
 
 # This pipeline will take all of the error votes, clean them, properly orders and formats them,
 # And outputs them into a friendly CSV format.
@@ -181,6 +148,5 @@ new_rep = (
         header=','.join(pol_full_lst)
     ))
 
+# Run the pipeline
 p.run()
-
-# If pipeline successful, then delete csv file
