+++ temp_poliviews/bill_votes_direct.py
@@ -44,7 +44,7 @@ topic_name = 'bill_votes'
 subscription_name = 'test_bill_votes_sub'
 
 # Locations in BigQuery that this script uses to read/write.
-dataset_id = 'poliviews'
+dataset_id = 'runner_pipelines'
 dest_table_id = 'all_votes' # Location for writing any voting information
 nickname_table_id = 'nicknames' # Location to read list of any nicknames in order to properly edit politician info.
